Applied copy edits to the content
In addition to editing the content, I removed the code that created the contents list and the indices and tables list at the bottom of the topic. Any navigation like the contents list should be provided in the left nav, and the "Index" and "Search Page" links did not work anyway.